% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtures.R
\name{complement}
\alias{complement}
\title{Complementary colours}
\usage{
complement(shades, space = NULL)
}
\arguments{
\item{shades}{One or more colours, in any suitable form (see
\code{\link{shade}}), or a palette function or scale.}

\item{space}{The space in which to take the complement.}
}
\value{
New colours of class \code{"shade"}, or a new palette function.
}
\description{
This function returns the complement of its argument, the "opposite" colours
in the specified space.
}
\examples{
complement("cyan")
complement("cyan", space="HSV")
complement("cyan", space="Lab")
}
