% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shade.R
\name{space}
\alias{space}
\title{Retrieve the space of a colour vector}
\usage{
space(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{...}{Additional arguments to methods.}
}
\value{
A string naming a colour space.
}
\description{
This function retrieves the colour space in which its argument is currently
defined.
}
\examples{
space("red")
}
\author{
Jon Clayden \href{mailto:code@clayden.org}{code@clayden.org}
}
