% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinytest.R
\name{expect_equal_shades}
\alias{expect_equal_shades}
\title{Expectation for comparing shades}
\usage{
expect_equal_shades(current, target, ..., info = NA_character_)
}
\arguments{
\item{current}{The colour vector to test.}

\item{target}{The target values to compare against, in any form valid for
casting to \code{\link[=shade]{shade()}}.}

\item{...}{Further arguments to \code{\link[=all.equal]{all.equal()}}.}

\item{info}{An optional information string shown in case of failure.}
}
\value{
A \code{"tinytest"} object.
}
\description{
This function provides an expectation for use with the \code{tinytest} package,
which compares two colour vectors via \link{all.equal.shade(hexonly=TRUE)}. It
evaluates \code{TRUE} if the vectors have the same length and resolve to the same
hex colour strings.
}
