% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shade.R
\name{shade}
\alias{shade}
\alias{shades}
\title{The shade class}
\usage{
shade(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{...}{Additional parameters (currently unused).}
}
\value{
A character vector of class \code{"shade"}, with additional
  attributes as follows.
    \item{space}{A string naming a color space.}
    \item{coords}{A matrix giving colour coordinates in the relevant space,
      one colour per row.}
}
\description{
Objects of class \code{"shade"} are simply standard R character vectors
representing one or more 8-bit (s)RGB colours in HTML-like hex format, but
with extra attributes for internal use.
}
\author{
Jon Clayden <code@clayden.org>
}

