% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shade.R
\name{space}
\alias{space}
\title{Retrieve the space of a colour vector}
\usage{
space(x, ...)
}
\arguments{
\item{An}{R object.}
}
\value{
A string naming a colour space.
}
\description{
This function retrieves the colour space in which its argument is currently
defined.
}
\examples{
space("red")
}
\author{
Jon Clayden <code@clayden.org>
}

