% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatch.R
\name{swatch}
\alias{swatch}
\title{Simple colour swatches}
\usage{
swatch(x, bg = "white", ...)
}
\arguments{
\item{x}{One or more colours, in any suitable form (see
\code{\link{shade}}).}

\item{bg}{A background colour.}

\item{...}{Additional arguments (currently unused).}
}
\description{
This function provides a simple visualisation of a colour series as a series
of boxes against the specified background colour. If the input has more than
one dimension then the boxes will be arranged in a grid (flattening further
dimensions after the second).
}
\examples{
swatch(c("red", "green", "blue"))
}
\author{
Jon Clayden \href{mailto:code@clayden.org}{code@clayden.org}
}
